Working, but non-tested support for non-master branches
This turned out to be a more complicated change than I expected.
I wanted backwards compatibility for the old config files, so needed to
use an interface when unmarshalling the JSON files.

The app runs and seems to do the right thing with different branches,
and with defaulting to master if no branch is specified. I think think
things would break if a branch name contains spaces or other punctuation
chars.

Instead of passing aound a path string, a {Path, Branch} struct is
passed. The only real place that needs to know the branch is the pattern
matching for the status, so the struct might be overkill.

The tests currently don't even compile.

TODO: Get tests compiling and passing - this might be tricky as I'm not
sure where to define the {Path, Branch} struct, as both the helpers in
internal/test_helpers/repo.go and the main code needs to use it, and
that looks circular.

TODO: Handle branches with spaces or other weird chars.

TODO: Handle the case that the local repos have the wrong branch checked
out.